+++ packages/kit/src/core/config/index.spec.js
@@ -1,6 +1,6 @@
-import { test } from 'uvu';
+import { test, suite } from 'uvu';
 import * as assert from 'uvu/assert';
-import { validate_config } from './index.js';
+import { validate_config, deep_merge } from './index.js';
 
 test('fills in defaults', () => {
 	const validated = validate_config({});
@@ -171,7 +171,19 @@ test('fails if paths.base is not root-relative', () => {
 				}
 			}
 		});
-	}, /^config\.kit\.paths\.base must be a root-relative path$/);
+	}, /^kit\.paths\.base option must be a root-relative path that starts but doesn't end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
+});
+
+test("fails if paths.base ends with '/'", () => {
+	assert.throws(() => {
+		validate_config({
+			kit: {
+				paths: {
+					base: '/github-pages/'
+				}
+			}
+		});
+	}, /^kit\.paths\.base option must be a root-relative path that starts but doesn't end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
 });
 
 test('fails if prerender.pages are invalid', () => {
@@ -286,3 +298,152 @@ validate_paths(
 );
 
 test.run();
+
+const deepMergeSuite = suite('deep_merge');
+
+deepMergeSuite('basic test no conflicts', async () => {
+	const [merged, conflicts] = deep_merge(
+		{
+			version: 1,
+			animalSounds: {
+				cow: 'moo'
+			}
+		},
+		{
+			animalSounds: {
+				duck: 'quack'
+			},
+			locale: 'en_US'
+		}
+	);
+	assert.equal(merged, {
+		version: 1,
+		locale: 'en_US',
+		animalSounds: {
+			cow: 'moo',
+			duck: 'quack'
+		}
+	});
+	assert.equal(conflicts, []);
+});
+
+deepMergeSuite('three way merge no conflicts', async () => {
+	const [merged, conflicts] = deep_merge(
+		{
+			animalSounds: {
+				cow: 'moo'
+			}
+		},
+		{
+			animalSounds: {
+				duck: 'quack'
+			}
+		},
+		{
+			animalSounds: {
+				dog: {
+					singular: 'bark',
+					plural: 'barks'
+				}
+			}
+		}
+	);
+	assert.equal(merged, {
+		animalSounds: {
+			cow: 'moo',
+			duck: 'quack',
+			dog: {
+				singular: 'bark',
+				plural: 'barks'
+			}
+		}
+	});
+	assert.equal(conflicts, []);
+});
+
+deepMergeSuite('merge with conflicts', async () => {
+	const [merged, conflicts] = deep_merge(
+		{
+			person: {
+				firstName: 'John',
+				lastName: 'Doe',
+				address: {
+					line1: '123 Main St',
+					city: 'Seattle',
+					state: 'WA'
+				}
+			}
+		},
+		{
+			person: {
+				middleInitial: 'Q',
+				address: '123 Main St, Seattle, WA'
+			}
+		}
+	);
+	assert.equal(merged, {
+		person: {
+			firstName: 'John',
+			middleInitial: 'Q',
+			lastName: 'Doe',
+			address: '123 Main St, Seattle, WA'
+		}
+	});
+	assert.equal(conflicts, ['person.address']);
+});
+
+deepMergeSuite('merge with arrays', async () => {
+	const [merged] = deep_merge(
+		{
+			paths: ['/foo', '/bar']
+		},
+		{
+			paths: ['/alpha', '/beta']
+		}
+	);
+	assert.equal(merged, {
+		paths: ['/foo', '/bar', '/alpha', '/beta']
+	});
+});
+
+deepMergeSuite('empty', async () => {
+	const [merged] = deep_merge();
+	assert.equal(merged, {});
+});
+
+deepMergeSuite('mutability safety', () => {
+	const input1 = {
+		person: {
+			firstName: 'John',
+			lastName: 'Doe',
+			address: {
+				line1: '123 Main St',
+				city: 'Seattle'
+			}
+		}
+	};
+	const input2 = {
+		person: {
+			middleInitial: 'L',
+			lastName: 'Smith',
+			address: {
+				state: 'WA'
+			}
+		}
+	};
+	const snapshot1 = JSON.stringify(input1);
+	const snapshot2 = JSON.stringify(input2);
+
+	const [merged] = deep_merge(input1, input2);
+
+	// Mess with the result
+	merged.person.middleInitial = 'Z';
+	merged.person.address.zipCode = '98103';
+	merged.person = {};
+
+	// Make sure nothing in the inputs changed
+	assert.snapshot(snapshot1, JSON.stringify(input1));
+	assert.snapshot(snapshot2, JSON.stringify(input2));
+});
+
+deepMergeSuite.run();