+++ packages/adapter-node/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sveltejs/adapter-node
 
+## 1.0.0-next.26
+
+### Minor Changes
+
+- 9a7195b: Allow sirv to looks for precompiled gzip and brotli files by default
+
+### Patch Changes
+
+- 53f3322: Fix build when using TypeScript and there is a `tsconfig.json` with `target: 'es2019'` or earlier
+
+## 1.0.0-next.25
+
+### Patch Changes
+
+- 0b780a6: Bundle server-side app during adapt phase
+
 ## 1.0.0-next.24
 
 ### Patch Changes