+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sveltejs/kit
 
+## 1.0.0-next.115
+
+### Patch Changes
+
+- 523c3e2: Allow vite.alias to be an array
+- 6fd46d1: \* update vite-plugin-svelte to 1.0.0-next.11 and use its named export
+  - update vite to 2.3.7
+- dc56d3c: Fix navigation when `base` path is set and validate that option's value
+
+## 1.0.0-next.114
+
+### Patch Changes
+
+- 5aa64ab: fix: SSL for HMR websockets #844
+- fae75f1: add optional state parameter for goto function
+- fbd5f8a: package command can now transpile TypeScript files
+
 ## 1.0.0-next.113
 
 ### Patch Changes